+++ src/store/animation.ts
@@ -50,8 +50,8 @@ import {
 } from '../utils/commons'
 import { getNotDuplicatedName } from '../utils/relations'
 import { useHistoryStore } from './history'
-import { useEntities } from '/@/composables/entities'
-import { useItemSelectable } from '/@/composables/selectable'
+import { useEntities } from '/@/composables/stores/entities'
+import { useItemSelectable } from '/@/composables/stores/selectable'
 import { useAnimationFrameStore } from '/@/composables/stores/animationFrame'
 import { HistoryStore } from '/@/composables/stores/history'
 import { useKeyframeStates } from '/@/composables/stores/keyframeStates'
