+++ src/store/animationGraph.ts
@@ -21,9 +21,9 @@ import { computed } from 'vue'
 import { AnimationGraph, IdMap, toMap } from '../models'
 import { extractMap, toList } from '../utils/commons'
 import { useHistoryStore } from './history'
-import { useEntities } from '/@/composables/entities'
+import { useEntities } from '/@/composables/stores/entities'
 import { SelectOptions } from '/@/composables/modes/types'
-import { useItemSelectable } from '/@/composables/selectable'
+import { useItemSelectable } from '/@/composables/stores/selectable'
 import { HistoryStore } from '/@/composables/stores/history'
 import { fromEntityList, toEntityList } from '/@/models/entity'
 import { GraphNode, GraphNodes, GraphNodeType } from '/@/models/graphNode'
