+++ src/store/element.ts
@@ -20,9 +20,9 @@ Copyright (C) 2021, Tomoya Komiyama.
 import { computed } from 'vue'
 import { Actor, BElement, getActor, toMap } from '../models'
 import { useHistoryStore } from './history'
-import { useEntities } from '/@/composables/entities'
+import { useEntities } from '/@/composables/stores/entities'
 import { SelectOptions } from '/@/composables/modes/types'
-import { useItemSelectable } from '/@/composables/selectable'
+import { useItemSelectable } from '/@/composables/stores/selectable'
 import { HistoryStore } from '/@/composables/stores/history'
 import { fromEntityList, toEntityList } from '/@/models/entity'
 import { getPlainSvgTree, toBElement } from '/@/utils/elements'
